feat(container metrics): change timeRange parameter type from Int to Long for better precision in time handling refactor(container metrics): rename ContainerMetricsPort to CpuMemoryMetricsPort for clarity and update related references feat(trace analysis): implement TraceAnalysisService to analyze HTTP requests, status codes, and latency from traces chore(trace repository): add TraceCountPerMinute data class to represent trace counts per minute refactor(span model): add methods to check if a span is an HTTP request and to retrieve its status code for better trace analysis

+++ src/main/kotlin/xquare/app/xquareinfra/application/container/port/in/ContainerMetricUseCase.kt
@@ -10,14 +10,14 @@ interface ContainerMetricUseCase {
     fun getContainerHttpRequestPerMinute(
         deployId: UUID,
         environment: ContainerEnvironment,
-        timeRange: Int,
+        timeRange: Long,
         user: User
     ): Map<String, Map<String, String>>
 
     fun getContainerHttpStatusRequestPerMinute(
         deployId: UUID,
         environment: ContainerEnvironment,
-        timeRange: Int,
+        timeRange: Long,
         statusCode: Int,
         user: User
     ): Map<String, Map<String, String>>
@@ -26,7 +26,7 @@ interface ContainerMetricUseCase {
         deployId: UUID,
         environment: ContainerEnvironment,
         percent: Int,
-        timeRange: Int,
+        timeRange: Long,
         user: User
     ): Map<String, Map<String, String>>
